AG-11543 - Move reference documentation files to folders
