+++ documentation/ag-grid-docs/src/components/reference-documentation/components/ApiDocumentationWithQuery.tsx
@@ -2,8 +2,8 @@ import { useStore } from '@nanostores/react';
 import { $queryClient } from '@stores/queryClientStore';
 import { QueryClientProvider } from '@tanstack/react-query';
 
+import { useResolvedInterfaces } from '../utils/useResolvedInterfaces';
 import { ApiDocumentation, type ApiDocumentationProps } from './ReferenceDocumentation';
-import { useResolvedInterfaces } from './useResolvedInterfaces';
 
 type Props = Omit<ApiDocumentationProps, 'interfaceLookup'>;
