+++ documentation/ag-grid-docs/src/components/reference-documentation/components/InterfaceDocumentation.astro
@@ -1,8 +1,8 @@
 ---
 import { InterfaceDocumentation } from './InterfaceDocumentation';
 import { getFrameworkFromPath } from '@features/docs/utils/urlPaths';
-import { getInterfaceDocumentationModel } from './getInterfaceDocumentationModel';
-import { getOverrides } from './getOverrides';
+import { getInterfaceDocumentationModel } from '../utils/getInterfaceDocumentationModel';
+import { getOverrides } from '../utils/getOverrides';
 import { getJsonFile } from '@utils/pages';
 import Code from '@ag-website-shared/components/code/Code';
