+++ documentation/ag-grid-docs/src/components/reference-documentation/components/Property.tsx
@@ -4,35 +4,18 @@ import { Icon } from '@ag-website-shared/components/icon/Icon';
 import { trackApiDocumentation } from '@utils/analytics';
 import { urlWithPrefix } from '@utils/urlWithPrefix';
 import classnames from 'classnames';
-import { Fragment, type FunctionComponent, type ReactElement, useEffect, useRef, useState } from 'react';
+import { Fragment, type FunctionComponent, useEffect, useRef, useState } from 'react';
 
-import styles from './ApiReference.module.scss';
+import type { ChildDocEntry, Config, ICallSignature, InterfaceEntry } from '../types';
 import {
     convertMarkdown,
-    escapeGenericCode,
     formatJsDocString,
-    getLinkedType,
     getTypeUrl,
     inferType,
     removeDefaultValue,
-} from './documentation-helpers';
-import { formatJson, getInterfaceName } from './interface-helpers';
-import type {
-    ChildDocEntry,
-    Config,
-    DocProperties,
-    ICallSignature,
-    InterfaceEntry,
-    ObjectCode,
-    Properties,
-    SectionProps,
-} from './types';
-
-interface Props {
-    framework: Framework;
-    model: DocProperties;
-    config: Config;
-}
+} from '../utils/documentation-helpers';
+import { formatJson, getInterfaceName } from '../utils/interface-helpers';
+import styles from './ApiReference.module.scss';
 
 function getDisplayNameSplit({ name, definition }: { name: string; definition: ChildDocEntry }) {
     let displayName = name;
@@ -170,68 +153,7 @@ function getTagsData({ definition, gridOpProp }: { definition: ChildDocEntry; gr
     };
 }
 
-const ObjectCodeSample: React.FC<ObjectCode> = ({ framework, id, breadcrumbs = {}, properties }) => {
-    const lines = [];
-    let indentationLevel = 0;
-
-    const getIndent = (level: number) => '  '.repeat(level);
-
-    Object.keys(breadcrumbs).forEach((key) => {
-        const indent = getIndent(indentationLevel);
-
-        if (indentationLevel > 0) {
-            lines.push(`${indent}...`);
-        }
-
-        lines.push(`${indent}${key}: {`);
-
-        indentationLevel++;
-    });
-
-    Object.entries(properties).forEach(([key, definition]) => {
-        let line = getIndent(indentationLevel) + key;
-
-        // process property object
-        if (!definition.isRequired) {
-            line += '?';
-        }
-
-        let type;
-        let isObject = false;
-
-        if (definition.meta && definition.meta.type != null) {
-            type = definition.meta.type;
-        } else if (definition.type != null) {
-            type = typeof definition.type === 'object' ? 'Function' : definition.type;
-        } else if (definition.options != null) {
-            type = definition.options.map((option) => formatJson(option)).join(' | ');
-        } else if (definition.default != null) {
-            type = Array.isArray(definition.default) ? 'object[]' : typeof definition.default;
-        } else if (definition.description != null) {
-            type = 'object';
-        } else {
-            type = getInterfaceName(key);
-            isObject = true;
-        }
-
-        line += `: ${isObject ? `<a href='#reference-${id}.${key}'>${type}</a>` : getLinkedType(type, framework)};`;
-
-        if (definition.default != null) {
-            line += ` // default: ${formatJson(definition.default)}`;
-        }
-
-        lines.push(line);
-    });
-
-    while (indentationLevel > 0) {
-        lines.push(`${getIndent(indentationLevel-- - 1)}}`);
-    }
-
-    const escapedLines = escapeGenericCode(lines);
-    return <Code code={escapedLines} keepMarkup={true} />;
-};
-
-const Property: FunctionComponent<{
+export const Property: FunctionComponent<{
     id: string;
     name: string;
     framework: Framework;
@@ -402,164 +324,4 @@ const Property: FunctionComponent<{
             )}
         </>
     );
-};
-
-const Breadcrumbs = ({ breadcrumbs }: { breadcrumbs: Record<string, string> }) => {
-    const breadcrumbsLength = Object.keys(breadcrumbs).length;
-
-    if (breadcrumbsLength <= 1) {
-        return null;
-    }
-
-    const links: ReactElement[] = [];
-    let href = '';
-    let index = 0;
-
-    Object.entries(breadcrumbs).forEach(([key, text]) => {
-        href += `${href.length > 0 ? '.' : 'reference-'}${key}`;
-
-        if (index < breadcrumbsLength - 1) {
-            links.push(
-                <Fragment key={key}>
-                    <a href={`#${href}`} title={text}>
-                        {key}
-                    </a>{' '}
-                    &gt;{' '}
-                </Fragment>
-            );
-        } else {
-            links.push(<Fragment key={key}>{key}</Fragment>);
-        }
-
-        index++;
-    });
-
-    return <div className={styles.breadcrumbs}>{links}</div>;
-};
-
-const SectionHeader = ({
-    title,
-    description,
-    page,
-    framework,
-    breadcrumbs = {},
-    displayName,
-    headerLevel,
-    hideHeader,
-    showSnippets,
-    properties,
-}: {
-    title: string;
-    description?: string;
-    page?: any;
-    framework: Framework;
-    breadcrumbs: Record<string, string>;
-    displayName: string;
-    headerLevel?: number;
-    hideHeader?: boolean;
-    showSnippets?: boolean;
-    properties: Properties;
-}) => {
-    breadcrumbs[title] = displayName;
-    const breadcrumbKeys = Object.keys(breadcrumbs);
-    const id = breadcrumbKeys.join('.');
-    const headerTagLevel = headerLevel || breadcrumbKeys.length + 1;
-    const HeaderTag = `h${headerTagLevel}` as any;
-    const descriptionDisplay = convertMarkdown(description, framework);
-
-    return (
-        <>
-            {!hideHeader && (
-                <HeaderTag id={`reference-${id}`} style={{ position: 'relative' }}>
-                    {displayName}
-                    <a href={`#reference-${id}`} className="docs-header-icon">
-                        <Icon name="link" />
-                    </a>
-                </HeaderTag>
-            )}
-            <Breadcrumbs breadcrumbs={breadcrumbs} />
-            {descriptionDisplay && <p dangerouslySetInnerHTML={{ __html: descriptionDisplay }}></p>}
-            {page && (
-                <p>
-                    See <a href={urlWithPrefix({ url: page.url, framework })}>{page.name}</a> for more information.
-                </p>
-            )}
-            {showSnippets && (
-                <ObjectCodeSample framework={framework} id={id} breadcrumbs={breadcrumbs} properties={properties} />
-            )}
-        </>
-    );
-};
-
-const Section: FunctionComponent<SectionProps> = ({
-    title,
-    framework,
-    names = [],
-    properties,
-    config = {} as Config,
-    breadcrumbs = {},
-    meta,
-}) => {
-    const showHeader = !config.isSubset;
-    const breadcrumbKeys = Object.keys(breadcrumbs);
-    const id = breadcrumbKeys.join('.');
-    const displayName = meta?.displayName || title;
-
-    return (
-        <div className={styles.apiReferenceOuter}>
-            {showHeader && (
-                <SectionHeader
-                    title={title}
-                    description={meta.description}
-                    page={meta.page}
-                    framework={framework}
-                    breadcrumbs={breadcrumbs}
-                    displayName={displayName}
-                    headerLevel={config.headerLevel}
-                    hideHeader={config.hideHeader ?? true}
-                    showSnippets={config.showSnippets && names.length < 1}
-                    properties={properties}
-                />
-            )}
-            <table
-                className={classnames(styles.reference, styles.apiReference, 'no-zebra')}
-                style={config.overrideBottomMargin ? { marginBottom: config.overrideBottomMargin } : {}}
-            >
-                <colgroup>
-                    <col></col>
-                    <col></col>
-                </colgroup>
-                <tbody>
-                    {Object.entries(properties).map(([name, { definition, gridOpProp, detailsCode, propertyType }]) => {
-                        return (
-                            <Property
-                                key={name}
-                                id={id}
-                                name={name}
-                                framework={framework}
-                                definition={definition}
-                                gridOpProp={gridOpProp}
-                                detailsCode={detailsCode}
-                                propertyType={propertyType}
-                                config={config}
-                            />
-                        );
-                    })}
-                </tbody>
-            </table>
-        </div>
-    );
-};
-
-export const InterfaceDocumentation: FunctionComponent<Props> = ({ framework, model, config }) => {
-    return Object.entries(model.properties).map(([key, properties]) => (
-        <Section
-            key={key}
-            framework={framework}
-            title={key}
-            properties={properties}
-            config={config}
-            meta={model.meta}
-        />
-    ));
 };
