+++ documentation/ag-grid-docs/src/components/reference-documentation/utils/getPropertiesFromSource.ts
